# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Vibe Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface CreateUserRequest {
  username: string;
  password: string;
  role: UserRole;
}