# lib/api-zod/src/generated/types/userEnvelope.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Vibe Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface UserEnvelope {
  user: User;
}